move formutils factory methods into formutils
fix codacy util class naming
