+++ android-json-form-wizard/src/main/java/com/vijay/jsonwizard/activities/NoLocaleJsonWizardFormActivity.java
@@ -1,11 +1,10 @@
 package com.vijay.jsonwizard.activities;
 
 import com.vijay.jsonwizard.utils.FormUtils;
-import com.vijay.jsonwizard.utils.FormUtilsFactory;
 
 public class NoLocaleJsonWizardFormActivity extends JsonWizardFormActivity {
     @Override
     protected FormUtils getFormUtils() {
-        return FormUtilsFactory.noLocaleInstance();
+        return FormUtils.newInstanceWithNoLocale();
     }
 }
