--- android-json-form-wizard/src/main/java/com/vijay/jsonwizard/utils/FormUtilsFactory.java
@@ -1,18 +1,0 @@
-package com.vijay.jsonwizard.utils;
-
-import android.content.Context;
-
-public class FormUtilsFactory {
-    public static FormUtils newInstance(){
-        return new FormUtils();
-    }
-
-    public static FormUtils noLocaleInstance(){
-        return new FormUtils(){
-            @Override
-            protected String getLocaleFormIdentity(Context context, String formIdentity) {
-                return formIdentity;
-            }
-        };
-    }
-}
